Extract JUnit Statement as top level class
Extract JUnit Statement from EasyMockRunner as a top-level class for
re-use in EasyMockRule
