+++ easymock/src/main/java/org/easymock/EasyMockRule.java
@@ -15,6 +15,7 @@
  */
 package org.easymock;
 
+import org.easymock.internal.EasyMockStatement;
 import org.junit.rules.TestRule;
 import org.junit.runner.Description;
 import org.junit.runners.model.Statement;
